Improve shaders using the complex exponential function
The exponential function in the complex plane basically furls a grid
around the origin while keeping the aspect ratio of all squares in the
grid. We can just apply this function to the vertical cross-section
between the vertex and the player. This will make the planet look
actually spherical and reduce distortions.